+++ src/main/java/gregtech/common/metatileentities/storage/MetaTileEntityCreativeTank.java
@@ -84,7 +84,7 @@ public class MetaTileEntityCreativeTank extends MetaTileEntityQuantumTank {
 
     @Override
     public ModularPanel buildUI(PosGuiData guiData, PanelSyncManager guiSyncManager) {
-        return GTGuis.createQuantumPanel(this);
+        return GTGuis.createPanel(this, 176, 166);
     }
 
     @Override
