+++ src/main/java/gregtech/common/metatileentities/storage/MetaTileEntityQuantumStorage.java
@@ -1,11 +1,14 @@
 package gregtech.common.metatileentities.storage;
 
 import gregtech.api.capability.GregtechDataCodes;
+import gregtech.api.capability.IActiveOutputSide;
 import gregtech.api.capability.IQuantumController;
 import gregtech.api.capability.IQuantumStorage;
 import gregtech.api.gui.GuiTextures;
 import gregtech.api.gui.widgets.ClickButtonWidget;
 import gregtech.api.metatileentity.MetaTileEntity;
+import gregtech.api.mui.GTGuiTextures;
+import gregtech.api.mui.GTGuis;
 import gregtech.api.util.GTUtility;
 import gregtech.client.renderer.handler.BlockPosHighlightRenderer;
 import gregtech.client.renderer.texture.Textures;
@@ -18,30 +21,61 @@ import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.network.PacketBuffer;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import net.minecraftforge.items.IItemHandlerModifiable;
 
 import codechicken.lib.render.CCRenderState;
 import codechicken.lib.render.pipeline.IVertexOperation;
 import codechicken.lib.vec.Matrix4;
+import com.cleanroommc.modularui.api.drawable.IKey;
+import com.cleanroommc.modularui.factory.PosGuiData;
+import com.cleanroommc.modularui.screen.ModularPanel;
+import com.cleanroommc.modularui.utils.Alignment;
+import com.cleanroommc.modularui.utils.Color;
+import com.cleanroommc.modularui.value.sync.BooleanSyncValue;
+import com.cleanroommc.modularui.value.sync.PanelSyncManager;
+import com.cleanroommc.modularui.value.sync.SyncHandlers;
+import com.cleanroommc.modularui.widget.ParentWidget;
+import com.cleanroommc.modularui.widgets.ItemSlot;
+import com.cleanroommc.modularui.widgets.SlotGroupWidget;
+import com.cleanroommc.modularui.widgets.TextWidget;
+import com.cleanroommc.modularui.widgets.ToggleButton;
+import com.cleanroommc.modularui.widgets.layout.Column;
+import com.cleanroommc.modularui.widgets.layout.Row;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.ref.WeakReference;
 import java.util.List;
+import java.util.function.BooleanSupplier;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
-public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity implements IQuantumStorage<T> {
+import static gregtech.api.capability.GregtechDataCodes.*;
+
+public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity implements IQuantumStorage<T>,
+                                                  IActiveOutputSide {
 
     /** not synced, server only. lazily initialized from pos */
     private WeakReference<IQuantumController> controller = new WeakReference<>(null);
 
     /** synced, server and client */
     private BlockPos controllerPos;
+    protected static final String IS_VOIDING = "IsVoiding";
+    protected static final String INPUT_FROM_OUTPUT = "AllowInputFromOutputSide";
+    protected static final String INPUT_FROM_OUTPUT_FLUID = "AllowInputFromOutputSideF";
 
     private ClickButtonWidget connectedIcon;
+    protected EnumFacing outputFacing;
+    protected boolean voiding = false;
+    protected boolean autoOutput;
+    protected boolean allowInputFromOutputSide = false;
+    protected boolean locked = false;
 
     public MetaTileEntityQuantumStorage(ResourceLocation metaTileEntityId) {
         super(metaTileEntityId);
@@ -127,12 +161,171 @@ public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity imp
     }
 
     @Override
+    public boolean usesMui2() {
+        return true;
+    }
+
+    @Override
+    public ModularPanel buildUI(PosGuiData guiData, PanelSyncManager guiSyncManager) {
+        var panel = GTGuis.createPanel(this, 176, 166);
+        createWidgets(panel, guiSyncManager);
+        return panel.padding(4)
+                .child(IKey.lang(getMetaFullName()).asWidget())
+                .child(createQuantumIO(importItems, exportItems))
+                .child(createQuantumButtonRow())
+                .child(SlotGroupWidget.playerInventory().left(7));
+    }
+
+    protected void createWidgets(ModularPanel mainPanel, PanelSyncManager syncManager) {}
+
+    public Column createQuantumDisplay(String lang,
+                                       Supplier<String> name, Predicate<TextWidget> condition,
+                                       Supplier<String> count) {
+        return new Column()
+                .background(GTGuiTextures.DISPLAY)
+                .padding(4)
+                .height(46)
+                .top(16)
+                .child(IKey.lang(lang)
+                        .alignment(Alignment.TopLeft)
+                        .color(Color.WHITE.main)
+                        .asWidget()
+                        .widthRel(0.5f)
+                        .left(4)
+                        .marginBottom(2))
+                .child(IKey.dynamic(name)
+                        .alignment(Alignment.TopLeft)
+                        .color(Color.WHITE.main)
+                        .asWidget()
+                        .setEnabledIf(condition)
+                        .widthRel(0.5f)
+                        .left(4)
+                        .height(20)
+                        .marginBottom(2))
+                .child(IKey.dynamic(count)
+                        .alignment(Alignment.TopLeft)
+                        .color(Color.WHITE.main)
+                        .asWidget()
+                        .widthRel(0.5f)
+                        .left(4));
+    }
+
+    public ParentWidget<?> createQuantumIO(IItemHandlerModifiable importHandler,
+                                           IItemHandlerModifiable exportHandler) {
+        return new Row()
+                .pos(79, 18 + 45)
+                .coverChildren()
+                .child(new ItemSlot()
+                        .background(GTGuiTextures.SLOT, GTGuiTextures.IN_SLOT_OVERLAY)
+                        .slot(SyncHandlers.itemSlot(importHandler, 0)
+                                .accessibility(true, false)
+                                .singletonSlotGroup(200))
+                        .marginRight(18))
+                .child(new ItemSlot()
+                        .background(GTGuiTextures.SLOT, GTGuiTextures.OUT_SLOT_OVERLAY)
+                        .slot(SyncHandlers.itemSlot(exportHandler, 0)
+                                .accessibility(false, true)));
+    }
+
+    public Row createQuantumButtonRow() {
+        boolean isFluid = getType() == Type.FLUID;
+        BooleanSupplier getter = isFluid ? this::isAutoOutputFluids : this::isAutoOutputItems;
+
+        return new Row()
+                .coverChildren()
+                .pos(7, 63)
+                .child(new ToggleButton()
+                        .overlay(isFluid ? GTGuiTextures.BUTTON_FLUID_OUTPUT : GTGuiTextures.BUTTON_ITEM_OUTPUT)
+                        .value(new BooleanSyncValue(getter, this::setAutoOutput)))
+                .child(new ToggleButton()
+                        .overlay(GTGuiTextures.FLUID_LOCK_OVERLAY)
+                        .value(new BooleanSyncValue(this::isLocked, this::setLocked)))
+                .child(new ToggleButton()
+                        .overlay(isFluid ? GTGuiTextures.FLUID_VOID_OVERLAY : GTGuiTextures.ITEM_VOID_OVERLAY)
+                        .value(new BooleanSyncValue(this::isVoiding, this::setVoiding)));
+    }
+
+    protected boolean isVoiding() {
+        return this.voiding;
+    }
+
+    protected void setVoiding(boolean isVoiding) {
+        this.voiding = isVoiding;
+        if (!getWorld().isRemote) {
+            writeCustomData(UPDATE_IS_VOIDING, buf -> buf.writeBoolean(this.voiding));
+            markDirty();
+        }
+    }
+
+    protected boolean isLocked() {
+        return this.locked;
+    }
+
+    protected void setLocked(boolean locked) {
+        if (this.locked == locked) return;
+        this.locked = locked;
+        if (!getWorld().isRemote) {
+            markDirty();
+        }
+    }
+
+    public EnumFacing getOutputFacing() {
+        return outputFacing == null ? frontFacing.getOpposite() : outputFacing;
+    }
+
+    public void setOutputFacing(EnumFacing outputFacing) {
+        this.outputFacing = outputFacing;
+        if (!getWorld().isRemote) {
+            notifyBlockUpdate();
+            writeCustomData(UPDATE_OUTPUT_FACING, buf -> buf.writeByte(outputFacing.getIndex()));
+            markDirty();
+        }
+    }
+
+    public boolean isAutoOutputItems() {
+        return getType() == Type.ITEM && autoOutput;
+    }
+
+    public boolean isAutoOutputFluids() {
+        return getType() == Type.FLUID && autoOutput;
+    }
+
+    public void setAutoOutput(boolean autoOutput) {
+        this.autoOutput = autoOutput;
+        if (!getWorld().isRemote) {
+            writeCustomData(UPDATE_AUTO_OUTPUT_ITEMS, buf -> buf.writeBoolean(autoOutput));
+            markDirty();
+        }
+    }
+
+    @Override
+    public boolean isAllowInputFromOutputSideItems() {
+        return getType() == Type.ITEM && allowInputFromOutputSide;
+    }
+
+    @Override
+    public boolean isAllowInputFromOutputSideFluids() {
+        return getType() == Type.FLUID && allowInputFromOutputSide;
+    }
+
+    public void setAllowInputFromOutputSide(boolean allowInputFromOutputSide) {
+        this.allowInputFromOutputSide = allowInputFromOutputSide;
+        if (!getWorld().isRemote) {
+            markDirty();
+        }
+    }
+
+    @Override
     public void writeInitialSyncData(@NotNull PacketBuffer buf) {
         super.writeInitialSyncData(buf);
         buf.writeBoolean(controllerPos != null);
         if (controllerPos != null) {
             buf.writeBlockPos(controllerPos);
         }
+        buf.writeByte(getOutputFacing().getIndex());
+        buf.writeBoolean(autoOutput);
+        buf.writeBoolean(voiding);
+        buf.writeBoolean(locked);
     }
 
     @Override
@@ -142,6 +335,10 @@ public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity imp
             controllerPos = buf.readBlockPos();
             scheduleRenderUpdate();
         }
+        this.outputFacing = EnumFacing.VALUES[buf.readByte()];
+        this.autoOutput = buf.readBoolean();
+        this.voiding = buf.readBoolean();
+        this.locked = buf.readBoolean();
     }
 
     @Override
@@ -169,6 +366,14 @@ public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity imp
             // tell controller to highlight
             BlockPosHighlightRenderer.renderBlockBoxHighLight(getControllerPos(), 6000, 1500);
             Minecraft.getMinecraft().player.closeScreen();
+        } else if (dataId == UPDATE_OUTPUT_FACING) {
+            this.outputFacing = EnumFacing.VALUES[buf.readByte()];
+            scheduleRenderUpdate();
+        } else if (dataId == UPDATE_AUTO_OUTPUT_FLUIDS || dataId == UPDATE_AUTO_OUTPUT_ITEMS) {
+            this.autoOutput = buf.readBoolean();
+            scheduleRenderUpdate();
+        } else if (dataId == UPDATE_IS_VOIDING) {
+            setVoiding(buf.readBoolean());
         }
     }
 
@@ -179,6 +384,13 @@ public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity imp
         if (controllerPos != null) {
             tagCompound.setLong("ControllerPos", controllerPos.toLong());
         }
+        if (getType() == Type.ITEM || getType() == Type.FLUID) {
+            data.setInteger("OutputFacing", getOutputFacing().getIndex());
+            data.setBoolean("AutoOutputItems", autoOutput);
+            data.setBoolean(INPUT_FROM_OUTPUT, this.allowInputFromOutputSide);
+            data.setBoolean(IS_VOIDING, isVoiding());
+            data.setBoolean("IsLocked", locked);
+        }
         return tagCompound;
     }
 
@@ -187,6 +399,18 @@ public abstract class MetaTileEntityQuantumStorage<T> extends MetaTileEntity imp
         super.readFromNBT(data);
         if (data.getBoolean("HasController")) {
             this.controllerPos = BlockPos.fromLong(data.getLong("ControllerPos"));
+        }
+        if (getType() == Type.ITEM || getType() == Type.FLUID) {
+            this.outputFacing = EnumFacing.VALUES[data.getInteger("OutputFacing")];
+            this.autoOutput = data.getBoolean("AutoOutputItems");
+
+            if (data.hasKey(INPUT_FROM_OUTPUT))
+                this.allowInputFromOutputSide = data.getBoolean(INPUT_FROM_OUTPUT);
+            else if (data.hasKey(INPUT_FROM_OUTPUT_FLUID))
+                this.allowInputFromOutputSide = data.getBoolean(INPUT_FROM_OUTPUT_FLUID);
+
+            this.voiding = data.getBoolean(IS_VOIDING) || data.getBoolean("IsPartiallyVoiding"); // legacy save support
+            this.locked = data.getBoolean("IsLocked");
         }
     }
